requirements.txt
Django>=5.0,<5.1
GitPython
aiofile
llama-index
llama-index-embeddings-huggingface
llama-index-llms-ollama
python-dotenv
rich
sphinx
